pyproject.toml
[tool.poetry]
name = "miniproject"
version = "0.1.0"
description = ""
authors = ["Daniyal <daniyal.mambetov.04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
django-rest-framework = "^0.1.0"
jwt = "^1.3.1"
django-redis = "^5.4.0"
redis = "^5.2.1"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
pillow = "^11.1.0"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.5.0"
django-cors-headers = "^4.7.0"
django-taggit = "^6.1.0"
django-filter = "^25.1"
requests = "^2.32.3"
django-extensions = "^3.2.3"
packaging = "^24.2"
setuptools = "^75.8.2"
drf-yasg = "^1.21.9"
channels = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
